Add spec for simulation worker
